Moved validator to controller package
