+++ src/java/uts/isd/controller/LoginServlet.java
@@ -1,11 +1,9 @@
 package uts.isd.controller;
 
 import java.io.IOException;
-import java.sql.Connection;
 import java.sql.SQLException;
 
 import uts.isd.model.Account;
-import uts.isd.model.Validator;
 import uts.isd.model.dao.AccountDAO;
 import uts.isd.model.dao.CustomerDAO;
 import uts.isd.model.dao.DAOException;
